+++ head/link_test.go
@@ -1,9 +1,12 @@
 package head_test
 
 import (
+	"encoding/json"
 	"testing"
 
 	"github.com/invopop/gobl/head"
+	"github.com/invopop/gobl/pkg/here"
+	"github.com/invopop/jsonschema"
 	"github.com/invopop/validation"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -46,13 +49,23 @@ func TestLinkValidation(t *testing.T) {
 
 func TestLinkByKey(t *testing.T) {
 	t.Run("find link", func(t *testing.T) {
-		l1 := &head.Link{Key: "test1"}
-		l2 := &head.Link{Key: "test2"}
-		l3 := &head.Link{Key: "test3"}
+		l1 := &head.Link{Category: head.LinkCategoryKeyFormat, Key: "test1"}
+		l2 := &head.Link{Category: head.LinkCategoryKeyFormat, Key: "test2"}
+		l3 := &head.Link{Category: head.LinkCategoryKeyFormat, Key: "test3"}
 		list := []*head.Link{l1, l2, l3}
 
-		assert.Equal(t, l2, head.LinkByKey(list, "test2"))
-		assert.Nil(t, head.LinkByKey(list, "test4"))
+		assert.Equal(t, l2, head.LinkByCategoryAndKey(list, head.LinkCategoryKeyFormat, "test2"))
+		assert.Nil(t, head.LinkByCategoryAndKey(list, head.LinkCategoryKeyFormat, "test4"))
+	})
+
+	t.Run("find link with different categories", func(t *testing.T) {
+		l1 := &head.Link{Category: head.LinkCategoryKeyFormat, Key: "test1"}
+		l2 := &head.Link{Category: head.LinkCategoryKeyRequest, Key: "test1"}
+		list := []*head.Link{l1, l2}
+
+		assert.Equal(t, l1, head.LinkByCategoryAndKey(list, head.LinkCategoryKeyFormat, "test1"))
+		assert.Equal(t, l2, head.LinkByCategoryAndKey(list, head.LinkCategoryKeyRequest, "test1"))
+		assert.Nil(t, head.LinkByCategoryAndKey(list, head.LinkCategoryKeyResponse, "test1"))
 	})
 }
 
@@ -101,4 +114,38 @@ func TestDetectDuplicateLink(t *testing.T) {
 
 		require.ErrorContains(t, err, "duplicate key 'test1'")
 	})
+
+	t.Run("detect duplicate in category", func(t *testing.T) {
+		l1 := &head.Link{Category: head.LinkCategoryKeyFormat, Key: "test1", URL: "https://example.com"}
+		l2 := &head.Link{Category: head.LinkCategoryKeyFormat, Key: "test1", URL: "https://example.com/2"}
+		list := []*head.Link{l1, l2}
+
+		err := validation.Validate(list, head.DetectDuplicateLinks)
+
+		require.ErrorContains(t, err, "duplicate category 'format' and key 'test1'")
+	})
+}
+
+func TestLinkExtendJSONSchemas(t *testing.T) {
+	base := here.Doc(`
+		{
+			"properties": {
+				"category": {
+					"$ref": "https://gobl.org/draft-0/cbc/key",
+					"title": "Category"
+				}
+			}
+		}
+	`)
+	js := new(jsonschema.Schema)
+	require.NoError(t, json.Unmarshal([]byte(base), js))
+	head.Link{}.JSONSchemaExtend(js)
+
+	prop, ok := js.Properties.Get("category")
+	assert.True(t, ok)
+	assert.Len(t, prop.OneOf, 7)
+	assert.Equal(t, head.LinkCategoryKeyFormat, prop.OneOf[0].Const)
+	assert.Equal(t, "Format", prop.OneOf[0].Title)
+	assert.Equal(t, head.LinkCategoryKeyPortal, prop.OneOf[1].Const)
+	assert.Equal(t, "Portal", prop.OneOf[1].Title)
 }